refactor(modals): remove unreachable dialog paths
